separate sleep procedure into own lib
